Add typescript definitions
