+++ types/api/handcash_connect_api_error.d.ts
@@ -1,7 +1,9 @@
-export = HandCashConnectApiError;
 declare class HandCashConnectApiError extends Error {
-    constructor(httpStatusCode: any, message: any, info: any);
-    httpStatusCode: any;
-    message: any;
-    info: any;
+   constructor(httpStatusCode: number, message: string, info: string);
+
+   httpStatusCode: number;
+   message: string;
+   info: string;
 }
+
+export = HandCashConnectApiError;
