+++ types/environments.d.ts
@@ -1,16 +1,11 @@
-export namespace prod {
-    const apiEndpoint: string;
-    const clientUrl: string;
+export declare class Environment {
+    apiEndpoint: string;
+    clientUrl: string;
+    trustholderEndpoint: string;
 }
-export namespace beta {
-    const apiEndpoint_1: string;
-    export { apiEndpoint_1 as apiEndpoint };
-    const clientUrl_1: string;
-    export { clientUrl_1 as clientUrl };
-}
-export namespace iae {
-    const apiEndpoint_2: string;
-    export { apiEndpoint_2 as apiEndpoint };
-    const clientUrl_2: string;
-    export { clientUrl_2 as clientUrl };
+
+export default class Environments {
+    static prod: Environment;
+    static beta: Environment;
+    static iae: Environment;
 }
