+++ types/permissions.d.ts
@@ -1,0 +1,11 @@
+declare enum Permissions {
+    Pay = 'PAY',
+    UserPublicProfile = 'USER_PUBLIC_PROFILE',
+    UserPrivateProfile = 'USER_PRIVATE_PROFILE',
+    Friends = 'FRIENDS',
+    Decryption = 'DECRYPTION',
+    SignData = 'SIGN_DATA',
+    ReadBalance = 'READ_BALANCE',
+}
+
+export = Permissions;
