+++ types/profile/index.d.ts
@@ -1,14 +1,22 @@
-export = Profile;
+import Entities = require("../entities");
+import HandCashConnectService = require("../api/handcash_connect_service");
+
 declare class Profile {
-    constructor(handCashConnectService: any);
-    handCashConnectService: any;
-    getCurrentProfile(): Promise<any>;
-    getPublicProfilesByHandle(handles: any): Promise<any>;
-    getFriends(): Promise<any>;
-    getPermissions(): Promise<any>;
-    getEncryptionKeypair(): Promise<{
-        publicKey: any;
-        privateKey: any;
-    }>;
-    signData(dataSignatureParameters: any): Promise<any>;
+   constructor(handCashConnectService: HandCashConnectService);
+
+   handCashConnectService: HandCashConnectService;
+
+   getCurrentProfile(): Promise<Entities.UserPublicProfile>;
+
+   getPublicProfilesByHandle(handles: string[]): Promise<Entities.UserPublicProfile[]>;
+
+   getFriends(): Promise<Entities.UserPublicProfile[]>;
+
+   getPermissions(): Promise<any>;
+
+   getEncryptionKeypair(): Promise<Entities.KeyPair>;
+
+   signData(dataSignatureParameters: Entities.DataSignatureParameters): Promise<Entities.DataSignature>;
 }
+
+export = Profile;
