+++ types/run_extension/handcash_owner.d.ts
@@ -1,12 +1,17 @@
-export = HandCashOwner;
+import Environment = require("../environments");
+
 declare class HandCashOwner {
-    static fromAuthToken(authToken: any, env?: {
-        apiEndpoint: string;
-        clientUrl: string;
-    }, appSecret?: string): import("./handcash_owner");
-    constructor(handCashConnectService: any);
-    handCashConnectService: any;
-    nextOwner(alias: any): Promise<any>;
-    sign(rawTransaction: any, inputParents: any, locks: any): Promise<any>;
-    getNftLocations(): Promise<any>;
+   static fromAuthToken(authToken: any, env?: Environment.Environment, appSecret?: string): HandCashOwner;
+
+   constructor(handCashConnectService: any);
+
+   handCashConnectService: any;
+
+   nextOwner(alias: string): Promise<string>;
+
+   sign(rawTransaction: string, inputParents: any[], locks: any[]): Promise<string>;
+
+   getNftLocations(): Promise<string[]>;
 }
+
+export = HandCashOwner;
