+++ types/run_extension/handcash_purse.d.ts
@@ -1,11 +1,16 @@
-export = HandCashPurse;
+import Environments = require("../environments");
+import HandCashConnectService = require("../api/handcash_connect_service");
+
 declare class HandCashPurse {
-    static fromAuthToken(authToken: any, env?: {
-        apiEndpoint: string;
-        clientUrl: string;
-    }, appSecret?: string): import("./handcash_purse");
-    constructor(handCashConnectService: any);
-    handCashConnectService: any;
-    pay(rawTx: any, parents: any): Promise<any>;
-    broadcast(rawTx: any): Promise<void>;
+   static fromAuthToken(authToken: string, env?: Environments.Environment, appSecret?: string): HandCashPurse;
+
+   constructor(handCashConnectService: HandCashConnectService);
+
+   handCashConnectService: HandCashConnectService;
+
+   pay(rawTx: string, parents: any[]): Promise<any>;
+
+   broadcast(rawTx: string): Promise<void>;
 }
+
+export = HandCashPurse;
