+++ types/wallet/index.d.ts
@@ -1,10 +1,21 @@
-export = Wallet;
+import Entities = require("../entities");
+import HandCashConnectService = require("../api/handcash_connect_service");
+
+
 declare class Wallet {
-    constructor(handCashConnectService: any);
-    handCashConnectService: any;
-    getSpendableBalance(currencyCode: any): Promise<any>;
-    getTotalBalance(): Promise<any>;
-    pay(paymentParameters: any): Promise<any>;
-    getPayment(transactionId: any): Promise<any>;
-    getExchangeRate(currencyCode: any): Promise<any>;
+   constructor(handCashConnectService: HandCashConnectService);
+
+   handCashConnectService: HandCashConnectService;
+
+   getSpendableBalance(currencyCode: Entities.CurrencyCode): Promise<Entities.SpendableBalance>;
+
+   getTotalBalance(): Promise<Entities.UserBalance>;
+
+   pay(paymentParameters: Entities.PaymentParameters): Promise<Entities.PaymentResult>;
+
+   getPayment(transactionId: string): Promise<Entities.PaymentResult>;
+
+   getExchangeRate(currencyCode: Entities.CurrencyCode): Promise<Entities.ExchangeRate>;
 }
+
+export = Wallet;
